requirements.txt
Flask==3.0.0
pandas==2.1.4
scikit-learn==1.3.2
PyPDF2==3.0.1
python-docx==1.1.0
Werkzeug==3.0.1
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow==10.1.0
PyMuPDF==1.23.8
gunicorn==21.2.0
